interface/requirements.txt
pyserial==3.4
pyserial-asyncio==0.4
windows-curses==2.0.1
ipdb
aioserial==1.1.3
wxPython